02_realtime_streaming/requirements.txt
pandas
psycopg2-binary
kafka-python
joblib
xgboost
scikit-learn
